detection-model/requirements.txt
git+https://github.com/huggingface/transformers.git
timm
matplotlib
Pillow
pytesseract